feat: :fire: handle gatsby bug disallowing abstraction of querys out for gatsby-node
